+++ src/env/env.rs
@@ -53,7 +53,7 @@ fn print_env(null: bool) {
     }
 }
 
-pub fn uumain(args: Vec<String>) -> isize {
+pub fn uumain(args: Vec<String>) -> i32 {
     let prog = args[0].as_slice();
 
     // to handle arguments the same way than GNU env, we can't use getopts
@@ -198,7 +198,7 @@ pub fn uumain(args: Vec<String>) -> isize {
         match Command::new(prog).args(args).stdin(InheritFd(0)).stdout(InheritFd(1)).stderr(InheritFd(2)).status() {
             Ok(exit) =>
                 return match exit {
-                    std::old_io::process::ExitStatus(s) => s,
+                    std::old_io::process::ExitStatus(s) => s as i32,
                     _ => 1
                 },
             Err(_) => return 1
